requirements.txt
boto3
psycopg2-binary
awscli-local
